New Chart Controller
Added a controller to filter the chart separately from the list table.
It will stil reset when clicking on a sortable link in the list table,
not sure if that's a good design decision or not.

Additionally, there's now an extra list table column that spits out a
singular page title or taxonomy name.